# public/dist/swagger.yaml
swagger: '2.0'
info:
  description: API Documentation
  version: 1.0.0
  title: FoodPhone.menu
host: 'localhost:8080'
basePath: /api
...
schemes:
  - http
paths:
...
  /rol:
    post:
      tags:
        - rol
      summary: Create new rol
      description: ""
      operationId: createRol
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: new rol
          required: false
          schema:
            $ref: "#/definitions/Rol"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Rol"
        "400":
          description: Invalid Rol
  /rol/{rolId}:
    get:
      tags:
        - rol
      summary: Get specific rol
      description: ""
      operationId: getRolById
      produces:
        - application/json
      parameters:
        - name: rolId
          in: path
          description: ID of rol
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Rol'
        '400':
          description: Invalid ID supplied
        '404':
          description: Rol not found
    put:
      tags:
        - rol
      summary: Update an existing rol
      description: ""
      operationId: updateRol
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: rolId
          in: path
          description: ID of rol
          required: true
          type: string
        - in: body
          name: body
          description: Update rol
          required: false
          schema:
            $ref: "#/definitions/Rol"
      responses:
        "405":
          description: Validation exception
        "404":
          description: Rol not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - rol
      summary: Delete rol by ID
      description: ""
      operationId: deleteRol
      produces:
        - application/json
      parameters:
        - name: rolId
          in: path
          description: ID of the rol to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Rol'
        '400':
          description: Invalid ID supplied
        '404':
          description: Rol not found
          
#menuCategory
...
  /user:
    post:
      tags:
        - user
      summary: Create new user
      description: ""
      operationId: createUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New User
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/User"
        "400":
          description: Invalid User
  /user/{userId}:
    get:
      tags:
        - user
      summary: Get specific user
      description: ""
      operationId: getUserById
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of User
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update an existing user
      description: ""
      operationId: updateUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          type: string
        - in: body
          name: body
          description: Update User
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: Validation exception
        "404":
          description: User not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - user
      summary: Delete user by ID
      description: ""
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of the User to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found

#questionType
...
  /questionType:
    post:
      tags:
        - questionType
      summary: Create new QuestionType
      description: ""
      operationId: createQuestionType
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New QuestionType
          required: false
          schema:
            $ref: "#/definitions/QuestionType"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/QuestionType"
        "400":
          description: Invalid QuestionType
  /questionType/{questionTypeId}:
    get:
      tags:
        - questionType
      summary: Get specific QuestionType
      description: ""
      operationId: getQuestionTypeById
      produces:
        - application/json
      parameters:
        - name: questionTypeId
          in: path
          description: ID of QuestionType
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/QuestionType'
        '400':
          description: Invalid ID supplied
        '404':
          description: QuestionType not found
    put:
      tags:
        - questionType
      summary: Update an existing QuestionType
      description: ""
      operationId: updateQuestionType
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: questionTypeId
          in: path
          description: ID of questionType
          required: true
          type: string
        - in: body
          name: body
          description: Update QuestionType
          required: false
          schema:
            $ref: "#/definitions/QuestionType"
      responses:
        "405":
          description: Validation exception
        "404":
          description: QuestionType not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - questionType
      summary: Delete QuestionType by ID
      description: ""
      operationId: deleteQuestionType
      produces:
        - application/json
      parameters:
        - name: questionTypeId
          in: path
          description: ID of the QuestionType to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/QuestionType'
        '400':
          description: Invalid ID supplied
        '404':
          description: QuestionType not found

#parameter
...
definitions:
...
  Rol:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      isActive:
        type: boolean

#menuCategory
...
  User:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      roles:
        type: array
        items:
          type: string
      addresses:
        type: array
        items:
          $ref: "#/definitions/Addresses"
      phones:
        type: array
        items:
          $ref: "#/definitions/Phones"
  Addresses:
    type: object
    properties:
      name:
        type: string
      address:
        type: string
      details:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      isActive:
        type: boolean
  Phones:
    type: object
    properties:
      name:
        type: string
      number:
        type: string      
      isActive:
        type: boolean

#parameters
...
  QuestionType:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      numberId:
        type: number
      isActive:
        type: boolean